Added an alert to tell the user I havent written the logic to edit the transaction other than the note field.
Im still not sure if I will impliment that logic, it could get messy if the user changes the amount of a transaction which happens in the past of has an occurence in the past as this will edit their current balance. Who knows we will see.

If the people want it