Redesigned the main toggles the user can see on the main page.
The transaction and goal list area now has a title to show the user what they are looking at. This title itself is a menu picker, the user can click the title to switch between viewing transactions or goals.
This title moved the date range picker to the right side of the screen, and removed the chart style picker and date range adjuster arrows. The chart style picker has been moved to the settings menu. The date range adjustment will now be handled with a swipe left or right in the transaction list. - This is not yet implimented