Refined deleting transactions.
Singular transactions and archived instances of a recurring transaction now show a bright red delete button, while active instances of a recurring transaction still shows the deleting options.
The user will see an alert when they delete either a singular transactions or archived instances of a recurring transaction asking if they are sure. A warning message is also displayed when the transaction is in the past as deleting this transaction might have an effect on the current balance.